% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMM.Cluster.R
\name{LMM.Cluster}
\alias{LMM.Cluster}
\title{poly(A) sites clustering based on computational algorithm}
\usage{
LMM.Cluster(polyA, cutoff, method)
}
\arguments{
\item{polyA}{A data.frame containing seqname, strand, poly(A) sites, and gene id.}

\item{cutoff}{cutoff from LMMcutoff function.}

\item{method}{polyA cluster methods including polyAclust,peaklu, QuantifyR and simple cluster.}
}
\value{
The polyA data.frame will be return cluster of PolyA site.
}
\description{
poly(A) sites clustering based on computational algorithm
}
