% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMMcutoff.R
\name{LMMcutoff}
\alias{LMMcutoff}
\title{cross point poly(A) sites using expectation maximization and laplace mixture model}
\usage{
LMMcutoff(data,GenomeSize,p.adjusted)
}
\arguments{
\item{data}{A data.frame containing seqname, strand, poly(A) sites, and gene id.}

\item{GenomeSize}{A genome size of a species recommended.}

\item{p.adjusted}{Adjusted pvalue (BH).}
}
\value{
The polyA data.frame will be return cross point between cleavage sites within or different PolyA site.
}
\description{
cross point poly(A) sites using expectation maximization and laplace mixture model
}
